Fixed back button on both multiplayer and singleplayer with show functions and the BackButton class respectively. Lobby player's type is now removed on disconnect.

+++ Assets/Scripts/UI/_LevelSelection/SceneLoader.cs
@@ -42,7 +42,7 @@ public class SceneLoader : MonoBehaviour {
         {
             float progress = Mathf.Clamp01(operation.progress / 0.9f);
             slider.value = progress;
-            text.text = progress * 100f + "%";
+            text.text = (int)(progress * 100) + "%";
             yield return null;
         }
         loadScreen.SetActive(false);
